Move Match classes to beets.autotag.match
