+++ beets/autotag/__init__.py
@@ -22,17 +22,17 @@ from importlib import import_module
 from beets.util.deprecation import deprecate_for_maintainers, deprecate_imports
 
 from .distance import Distance, distance, string_dist, track_distance
-from .hooks import (
-    AlbumInfo,
+from .hooks import AlbumInfo, AttrDict, Info, TrackInfo, correct_list_fields
+from .match import (
     AlbumMatch,
-    AttrDict,
-    Info,
     Match,
-    TrackInfo,
+    Proposal,
+    Recommendation,
     TrackMatch,
-    correct_list_fields,
+    assign_items,
+    tag_album,
+    tag_item,
 )
-from .match import Proposal, Recommendation, assign_items, tag_album, tag_item
 
 
 def __getattr__(name: str):
